[SVG] Cast -1 values to unsigned to indicate we really did mean to type them

+++ src/cairo-svg-surface.c
@@ -930,7 +930,7 @@ emit_composite_image_pattern (cairo_output_stream_t     *output,
 
     status = _cairo_pattern_acquire_surface ((cairo_pattern_t *)pattern,
 					     (cairo_surface_t *)svg_surface,
-					     0, 0, -1, -1,
+					     0, 0, (unsigned int)-1, (unsigned int)-1,
 					     &surface, &surface_attr);
     if (status)
 	return status;
